refactor(rollback): add public barrel; rollback-service is a re-export
Move the composition root into rollback/index.ts — it binds defaultRollbackDeps()
and re-exports the frozen public API (6 functions + 5 types). rollback-service.ts
is now a one-line re-export barrel. ai-undo-service and rollback-to-point-service
compile with zero import changes; the existing mock-heavy rollback-service.test.ts
(incl. the userScopedAccess false-conflict regression) and ai-undo-service.test.ts
pass unmodified. Full api service suite (438 tests) green; typecheck green.
